feat: Add an option to show unreconciled accounts in sidebar (#3576)
This adds a settings option to the `uncleared-account-highlight`
extension to highlight unreconciled accounts in addition to uncleared
accounts.

When applied, accounts with cleared transactions in the register will
display the "filled" cleared transaction sprite next to the account in
the sidebar. This is in addition to continuing to display the "unfilled"
cleared transaction sprite for accounts with uncleared transactions in
the register.